Support for using external json files for input parameters.
You may copy the sample json file from data/fea and put it where the bin is located inside a folder if you would like to have many of them.
Milad-AC-GaitCycle/1.json

in the run time, provide the input argument the address of the json file such as:

./bin/Milad-AC-GaitCycle Milad-AC-GaitCycle/1.json

This will create a Milad-AC-GaitCycle folder for each json input, inside which there are VTK folder and 2Dplots and a copy of the original json file.

Enjoy!